requirements.txt
torch~=2.8.0
torchvision~=0.23.0 
numpy~=2.3
pillow~=12.0
s3torchconnector~=1.4.3 
polars~=1.34 
diffusers==0.34.0 
boto3~=1.40 
wandb~=0.22.3 
pandas~=2.3
pyarrow~=22.0
imageio[ffmpeg]~=2.37